Chore: Update .gitignore
