+++ .gitignore
@@ -2,3 +2,4 @@ __pycache__
 data
 .todo.txt
 container
+script.py
